state 값에 따른 삭제 로직 수정

+++ board-backend-boot/src/main/java/com/joo/service/BoardService.java
@@ -53,5 +53,5 @@ public interface BoardService {
      * @param boardId
      * @return
      */
-    void deleteBoardById(int boardId);
+    void deleteBoardById(Long boardId);
 }
